Return a standardised set of labels for the classification

Return a standardised set of labels for the classification

class_labels(
  type = c("trend", "effect"),
  lang = c("en", "nl"),
  detailed = TRUE,
  signed = TRUE
)

Arguments

type

What type of effect. Currently available are "trend" and "effect".

lang

The language. Currently available are "en" (English) and "nl" (Dutch). Defaults to "en". Please contact the maintainer if you have suggestions for more languages.

detailed

TRUE indicates a detailed classification(); FALSE a coarse_classification(). Defaults to TRUE.

signed

TRUE indicates a signed classification; FALSE a classification with remove_sign(). Defaults to TRUE.

See also