Classify Effects by Comparing the Confidence Intervals with a Reference and Thresholds

++ strong positive effect: max(threshold) < lcl

+ positive effect: reference < lcl < max(threshold) and max(threshold) < ucl

+~ moderate positive effect: reference < lcl and ucl < max(threshold)

~ no effect: min(threshold) < lcl < reference and reference < ucl < max(threshold)

-~ moderate negative effect: min(threshold) < lcl and ucl < reference

- negative effect: lcl < min(threshold) and min(threshold) < ucl < reference

-- strong negative effect: ucl < min(threshold)

?+ potential positive effect: min(threshold) < lcl < reference and max(threshold) < ucl

?- potential negative effect: lcl < min(threshold) and reference < ucl < max(threshold)

? unknown effect: lcl < min(threshold) and max(threshold) < ucl

classification(lcl, ucl, threshold, reference = 0)

Arguments

lcl

A vector of lower confidence limits.

ucl

A vector of upper confidence limits.

threshold

A vector of either 1 or 2 thresholds. A single threshold will be transformed into reference + c(-abs(threshold), abs(threshold)).

reference

The null hypothesis. Defaults to 0.

See also